Remove unnecessary ResourceLoader
