+++ core/native/src/main/scala/eu/joaocosta/minart/backend/defaults/package.scala
@@ -12,6 +12,6 @@ package object defaults {
   implicit val defaultPlatform: DefaultBackend[Any, Platform.Native.type] =
     DefaultBackend.fromConstant(Platform.Native)
 
-  implicit val defaultResourceLoader: DefaultBackend[Any, NativeResourceLoader.type] =
-    DefaultBackend.fromConstant(NativeResourceLoader)
+  implicit val defaultResource: DefaultBackend[String, NativeResource] =
+    DefaultBackend.fromFunction[String, NativeResource](NativeResource.apply)
 }
